+++ internal/provider/cluster_resource_test.go
@@ -22,6 +22,7 @@ import (
 	"log"
 	"net/http"
 	"os"
+	"regexp"
 	"testing"
 
 	"github.com/cockroachdb/cockroach-cloud-sdk-go/pkg/client"
@@ -319,6 +320,7 @@ func serverlessClusterWithSpendLimit(clusterName string) resource.TestStep {
 			resource.TestCheckNoResourceAttr(dataSourceName, "serverless.spend_limit"),
 			resource.TestCheckResourceAttrSet(dataSourceName, "serverless.usage_limits.request_unit_limit"),
 			resource.TestCheckResourceAttrSet(dataSourceName, "serverless.usage_limits.storage_mib_limit"),
+			resource.TestCheckResourceAttr(dataSourceName, "delete_protection", "false"),
 		),
 	}
 }
@@ -562,6 +564,7 @@ func multiRegionServerlessClusterResourceRegionUpdate(clusterName string) resour
 			resource.TestCheckNoResourceAttr(dataSourceName, "serverless.spend_limit"),
 			resource.TestCheckResourceAttr(dataSourceName, "serverless.usage_limits.request_unit_limit", "10000000000"),
 			resource.TestCheckResourceAttr(dataSourceName, "serverless.usage_limits.storage_mib_limit", "102400"),
+			resource.TestCheckResourceAttr(dataSourceName, "delete_protection", "false"),
 		),
 	}
 }
@@ -618,9 +621,15 @@ func TestIntegrationDedicatedClusterResource(t *testing.T) {
 	finalizedCluster := pendingCluster
 	finalizedCluster.UpgradeStatus = client.CLUSTERUPGRADESTATUSTYPE_FINALIZED
 
-	scaledCluster := finalizedCluster
+	firstUpdateCluster := finalizedCluster
+	firstUpdateCluster.DeleteProtection = ptr(client.DELETEPROTECTIONSTATETYPE_ENABLED)
+
+	secondUpdateCluster := firstUpdateCluster
+	secondUpdateCluster.DeleteProtection = ptr(client.DELETEPROTECTIONSTATETYPE_DISABLED)
+
+	scaledCluster := secondUpdateCluster
 	scaledCluster.Config.Dedicated = &client.DedicatedHardwareConfig{}
-	*scaledCluster.Config.Dedicated = *finalizedCluster.Config.Dedicated
+	*scaledCluster.Config.Dedicated = *secondUpdateCluster.Config.Dedicated
 	scaledCluster.Config.Dedicated.NumVirtualCpus = 4
 
 	httpOk := &http.Response{Status: http.StatusText(http.StatusOK)}
@@ -653,7 +662,7 @@ func TestIntegrationDedicatedClusterResource(t *testing.T) {
 		)
 
 	s.EXPECT().GetCluster(gomock.Any(), clusterID).
-		Return(&upgradingCluster, httpOk, nil).Times(1)
+		Return(&upgradingCluster, httpOk, nil)
 
 	// Scale (no-op)
 
@@ -664,12 +673,9 @@ func TestIntegrationDedicatedClusterResource(t *testing.T) {
 		})
 
 	s.EXPECT().GetCluster(gomock.Any(), clusterID).
-		Return(&pendingCluster, httpOk, nil).Times(2)
+		Return(&pendingCluster, httpOk, nil).Times(3)
 
 	// Finalize
-
-	s.EXPECT().GetCluster(gomock.Any(), clusterID).
-		Return(&pendingCluster, httpOk, nil).Times(1)
 
 	s.EXPECT().UpdateCluster(gomock.Any(), clusterID, gomock.Any()).
 		DoAndReturn(func(context.Context, string, *client.UpdateClusterSpecification,
@@ -678,12 +684,34 @@ func TestIntegrationDedicatedClusterResource(t *testing.T) {
 		})
 
 	s.EXPECT().GetCluster(gomock.Any(), clusterID).
-		Return(&finalizedCluster, httpOk, nil).Times(3)
+		Return(&finalizedCluster, httpOk, nil).Times(6)
 
 	// Import state happens here
 
+	// First Update
+
+	s.EXPECT().UpdateCluster(gomock.Any(), clusterID, gomock.Any()).
+		DoAndReturn(func(context.Context, string, *client.UpdateClusterSpecification,
+		) (*client.Cluster, *http.Response, error) {
+			currentCluster := &firstUpdateCluster
+			return currentCluster, httpOk, nil
+		})
+
 	s.EXPECT().GetCluster(gomock.Any(), clusterID).
-		Return(&finalizedCluster, httpOk, nil).Times(3)
+		Return(&firstUpdateCluster, httpOk, nil).Times(7)
+
+	// Failed Delete Attempt
+
+	// Second Update
+	s.EXPECT().UpdateCluster(gomock.Any(), clusterID, gomock.Any()).
+		DoAndReturn(func(context.Context, string, *client.UpdateClusterSpecification,
+		) (*client.Cluster, *http.Response, error) {
+			currentCluster := &secondUpdateCluster
+			return currentCluster, httpOk, nil
+		})
+
+	s.EXPECT().GetCluster(gomock.Any(), clusterID).
+		Return(&secondUpdateCluster, httpOk, nil).Times(6)
 
 	// Scale
 
@@ -695,14 +723,14 @@ func TestIntegrationDedicatedClusterResource(t *testing.T) {
 		})
 
 	s.EXPECT().GetCluster(gomock.Any(), clusterID).
-		Return(&scaledCluster, httpOk, nil).Times(5)
+		Return(&scaledCluster, httpOk, nil).AnyTimes()
 
 	// Deletion
 
 	s.EXPECT().DeleteCluster(gomock.Any(), clusterID)
 
 	scaleStep := resource.TestStep{
-		Config: getTestDedicatedClusterResourceConfig(clusterName, latestClusterMajorVersion, false, 4),
+		Config: getTestDedicatedClusterResourceConfig(clusterName, latestClusterMajorVersion, false, 4, nil),
 		Check:  resource.TestCheckResourceAttr("cockroach_cluster.dedicated", "dedicated.num_virtual_cpus", "4"),
 	}
 
@@ -719,7 +747,7 @@ func testDedicatedClusterResource(
 
 	testSteps := []resource.TestStep{
 		{
-			Config: getTestDedicatedClusterResourceConfig(clusterName, minSupportedClusterMajorVersion, false, 2),
+			Config: getTestDedicatedClusterResourceConfig(clusterName, minSupportedClusterMajorVersion, false, 2, nil),
 			Check: resource.ComposeTestCheckFunc(
 				testCheckCockroachClusterExists(resourceName),
 				resource.TestCheckResourceAttr(resourceName, "name", clusterName),
@@ -733,7 +761,7 @@ func testDedicatedClusterResource(
 			),
 		},
 		{
-			Config: getTestDedicatedClusterResourceConfig(clusterName, latestClusterMajorVersion, true, 2),
+			Config: getTestDedicatedClusterResourceConfig(clusterName, latestClusterMajorVersion, true, 2, nil),
 			Check: resource.ComposeTestCheckFunc(
 				resource.TestCheckResourceAttr(resourceName, "cockroach_version", latestClusterMajorVersion),
 				resource.TestCheckResourceAttr(dataSourceName, "cockroach_version", latestClusterMajorVersion),
@@ -756,6 +784,20 @@ func testDedicatedClusterResource(
 			// remains so after so I'm unclear why this is failing.  For now its
 			// turned off due to this reason.
 			ImportStateVerify: false,
+		},
+		{
+			Config: getTestDedicatedClusterResourceConfig(clusterName, latestClusterMajorVersion, false, 2, ptr(true)),
+			Check:  resource.TestCheckResourceAttr(resourceName, "delete_protection", "true"),
+		},
+		{
+			// Delete step that fails since delete protection is enabled.
+			Config:      " ",
+			Destroy:     true,
+			ExpectError: regexp.MustCompile(".*Cannot destroy cluster with delete protection enabled*"),
+		},
+		{
+			Config: getTestDedicatedClusterResourceConfig(clusterName, latestClusterMajorVersion, false, 2, ptr(false)),
+			Check:  resource.TestCheckResourceAttr(resourceName, "delete_protection", "false"),
 		},
 	}
 	testSteps = append(testSteps, additionalSteps...)
@@ -792,7 +834,14 @@ func testCheckCockroachClusterExists(resourceName string) resource.TestCheckFunc
 	}
 }
 
-func getTestDedicatedClusterResourceConfig(name, version string, finalize bool, vcpus int) string {
+func getTestDedicatedClusterResourceConfig(
+	name, version string, finalize bool, vcpus int, deleteProtectionEnabled *bool,
+) string {
+	var deleteProtectionConfig string
+	if deleteProtectionEnabled != nil {
+		deleteProtectionConfig = fmt.Sprintf("\ndelete_protection = %t\n", *deleteProtectionEnabled)
+	}
+
 	config := fmt.Sprintf(`
 resource "cockroach_cluster" "dedicated" {
     name           = "%s"
@@ -806,12 +855,13 @@ resource "cockroach_cluster" "dedicated" {
 		name: "us-central1"
 		node_count: 1
 	}]
+	%s
 }
 
 data "cockroach_cluster" "test" {
     id = cockroach_cluster.dedicated.id
 }
-`, name, version, vcpus)
+`, name, version, vcpus, deleteProtectionConfig)
 
 	if finalize {
 		config += fmt.Sprintf(`
@@ -901,4 +951,8 @@ func TestClusterSchemaInSync(t *testing.T) {
 	rAttrs := rSchema.Schema.Attributes
 	dAttrs := dSchema.Schema.Attributes
 	CheckSchemaAttributesMatch(t, rAttrs, dAttrs)
+}
+
+func ptr[T any](in T) *T {
+	return &in
 }